Extract V2's runtime from the V1 driver, then remove V1
V1 is obsolete, but it was not dangling: every module of
examples/gabriels_workflow_v2/ imported from it, and V2's workflow subclassed
its AgentGateway and GitRepository. Deleting it outright would have taken V2
with it. So the half V2 depends on moves into V2 as leaf modules, and the
half that was only V1's decisions goes.

Moved into examples/gabriels_workflow_v2/:

- errors.py    WorkflowError/WorkflowStopped, LOGGER, configure_logging
- gates.py     GateResult, CommandResult, and reading each gate's verdict
               back out of one interleaved `make ci` log
- gateway.py   AgentGateway and the whole bwrap sandbox stack it builds
- git.py       GitRepository: branch, worktree, commit, push, run CI
- github_app.py  the GitHub App client, plus the markdown renderer and the
               PR-URL parser that used to live in development_workflow.py

The renderer moving is what lets github_app.py drop its function-local
`import _render_comment`: that import existed only to break the cycle between
two V1 modules that imported each other.

Three helpers had exactly one consumer left and moved there instead of into a
shared module: _outcome and _shorten_home into workflow.py, and
_pull_request_number into github_app.py as pull_request_number. Names that now
cross a module boundary lost their underscore (readable, bounded,
gate_results); the ones still private to their module kept it.

Deleted with V1: DevelopmentWorkflow, the comment-as-state GitHub class,
ArtifactStore, the V1 CLI, its prompts and validation schemas, its config and
setup, simple_development_workflow.py, and the helpers only they used
(_attribution, _ci_signature, _markdown's V1 callers, the context-bounding
family).

Tests split the same way rather than being deleted with the file:

- tests/test_agent_gateway.py    the gateway and sandbox argv unit tests
- tests/test_git_and_gates.py    GitRepository and gate parsing
- tests/test_github_app.py       the App client and comment rendering
- tests/test_development_workflow.py is gone; what it covered that no longer
  exists went with it.

The moved gateway tests now run against V2's real prompts and schemas, so
their placeholders are CONTEXT_JSON and their schema is proposal/work-report.
test_sandbox.py and test_quality_gates.py repoint to the new modules.

Every new module lands at 98%+ coverage and the changed lines at 100%, so the
90% floors move to the new paths without weakening the ratchet.

Dependencies: pydantic, PyYAML, jsonschema, and pygithub are all still
imported. PyJWT and cryptography are not, and never were — pygithub pulls
pyjwt[crypto] itself, which is what its App auth signs with — so they stop
being declared here.

`make ci` passes: all eleven gates, 650 tests at 99.26% coverage, mutation
98.8%.
