+++ SKILLS/addyosmani/code-review-and-quality/SKILL.md
@@ -9,8 +9,8 @@ Vendored from https://github.com/addyosmani/agent-skills/blob/main/skills/code-r
 "See Also" section below points at our own pipeline instead of sibling skills
 that don't exist here, and a mapping note ties the five axes and severity
 vocabulary below to the `axis` / `severity` fields the automated review agent
-emits under examples/gabriels_workflow/prompts/review.md and
-examples/gabriels_workflow/validations/review.json.
+emits under examples/gabriels_workflow_v2/prompts/review-quality.md
+and examples/gabriels_workflow_v2/validations/review.json.
 -->
 
 # Code Review and Quality
@@ -198,7 +198,7 @@ Label every comment with its severity so the author knows what's required vs opt
 
 This prevents authors from treating all feedback as mandatory and wasting time on optional suggestions.
 
-> **Automated pipeline note:** `examples/gabriels_workflow/validations/review.json` requires every finding to carry a non-empty `required_change`, so purely informational **FYI** notes have no home there — fold anything worth keeping into the review's free-text `summary` instead of inventing a finding with no action.
+> **Automated pipeline note:** `examples/gabriels_workflow_v2/validations/review.json` requires every finding to carry a non-empty `required_change`, so purely informational **FYI** notes have no home there — fold anything worth keeping into the review's free-text `summary` instead of inventing a finding with no action.
 
 **Lead with what matters.** Order findings by leverage: correctness and security first, then structural regressions and missed simplifications, then everything else. Don't bury a real issue under cosmetic nits — a few high-conviction comments beat a long list. If you have one structural problem and ten nits, the structural problem *is* the review.
 
@@ -363,8 +363,8 @@ This repo doesn't vendor Addy Osmani's sibling `security-and-hardening` / `perfo
 
 For how this skill's axes and severities feed the automated review loop, see:
 
-- `examples/gabriels_workflow/prompts/review.md` — the prompt that instructs the `quality` review agent to apply this checklist
-- `examples/gabriels_workflow/validations/review.json` — the JSON schema each finding (`axis`, `severity`, `title`, `evidence`, `required_change`) must satisfy, which is what lets a finding be rendered straight into a GitHub PR comment
+- `examples/gabriels_workflow_v2/prompts/review-quality.md` — the prompt that instructs the `quality` review agent to apply this checklist
+- `examples/gabriels_workflow_v2/validations/review.json` — the JSON schema each finding (`axis`, `severity`, `title`, `evidence`, `required_change`) must satisfy, which is what lets a finding be rendered straight into a GitHub PR comment
 
 ## Common Rationalizations
