--- examples/gabriels_workflow/__init__.py
@@ -1,1 +1,0 @@
-"""Gabriel's issue-to-pull-request development workflow."""
