--- examples/gabriels_workflow/validations/expansion.json
@@ -1,45 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "type": "object",
-  "additionalProperties": false,
-  "required": [
-    "decision",
-    "needs_another_round",
-    "reason",
-    "summary",
-    "current_state",
-    "proposed_changes",
-    "out_of_scope",
-    "risks",
-    "open_questions"
-  ],
-  "properties": {
-    "decision": {
-      "type": "string",
-      "enum": ["proceed", "stop"]
-    },
-    "needs_another_round": {"type": "boolean"},
-    "reason": {"type": "string", "minLength": 1},
-    "summary": {"type": "string", "minLength": 1},
-    "current_state": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    },
-    "proposed_changes": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    },
-    "out_of_scope": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    },
-    "risks": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    },
-    "open_questions": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    }
-  }
-}
