--- examples/gabriels_workflow/validations/implementation.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "type": "object",
-  "additionalProperties": false,
-  "required": ["status", "summary", "files_changed", "tests_run", "blockers"],
-  "properties": {
-    "status": {
-      "type": "string",
-      "enum": ["complete", "blocked"]
-    },
-    "summary": {"type": "string", "minLength": 1},
-    "files_changed": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    },
-    "tests_run": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    },
-    "blockers": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    }
-  }
-}
