--- examples/gabriels_workflow/validations/review.json
@@ -1,47 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "type": "object",
-  "additionalProperties": false,
-  "required": [
-    "verdict",
-    "needs_another_round",
-    "reason",
-    "summary",
-    "findings"
-  ],
-  "properties": {
-    "verdict": {
-      "type": "string",
-      "enum": ["approve", "changes_requested"]
-    },
-    "needs_another_round": {"type": "boolean"},
-    "reason": {"type": "string", "minLength": 1},
-    "summary": {"type": "string", "minLength": 1},
-    "findings": {
-      "type": "array",
-      "items": {
-        "type": "object",
-        "additionalProperties": false,
-        "required": ["severity", "axis", "title", "evidence", "required_change"],
-        "properties": {
-          "severity": {
-            "type": "string",
-            "enum": ["critical", "required", "optional", "nit"]
-          },
-          "axis": {
-            "type": "string",
-            "enum": [
-              "specification",
-              "correctness",
-              "readability",
-              "architecture"
-            ]
-          },
-          "title": {"type": "string", "minLength": 1},
-          "evidence": {"type": "string", "minLength": 1},
-          "required_change": {"type": "string", "minLength": 1}
-        }
-      }
-    }
-  }
-}
