--- examples/gabriels_workflow/validations/specification.json
@@ -1,44 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "type": "object",
-  "additionalProperties": false,
-  "required": [
-    "title",
-    "problem_statement",
-    "solution",
-    "user_stories",
-    "implementation_decisions",
-    "testing_decisions",
-    "acceptance_criteria",
-    "out_of_scope"
-  ],
-  "properties": {
-    "title": {"type": "string", "minLength": 1},
-    "problem_statement": {"type": "string", "minLength": 1},
-    "solution": {"type": "string", "minLength": 1},
-    "user_stories": {
-      "type": "array",
-      "minItems": 1,
-      "items": {"type": "string", "minLength": 1}
-    },
-    "implementation_decisions": {
-      "type": "array",
-      "minItems": 1,
-      "items": {"type": "string", "minLength": 1}
-    },
-    "testing_decisions": {
-      "type": "array",
-      "minItems": 1,
-      "items": {"type": "string", "minLength": 1}
-    },
-    "acceptance_criteria": {
-      "type": "array",
-      "minItems": 1,
-      "items": {"type": "string", "minLength": 1}
-    },
-    "out_of_scope": {
-      "type": "array",
-      "items": {"type": "string", "minLength": 1}
-    }
-  }
-}
